change: Restructuring Training Compiler UI implementation
