+++ src/sagemaker/huggingface/__init__.py
@@ -17,4 +17,4 @@ from sagemaker.huggingface.estimator import HuggingFace  # noqa: F401
 from sagemaker.huggingface.model import HuggingFaceModel, HuggingFacePredictor  # noqa: F401
 from sagemaker.huggingface.processing import HuggingFaceProcessor  # noqa:F401
 
-from sagemaker.training_compiler.config import TrainingCompilerConfig  # noqa: F401
+from sagemaker.huggingface.training_compiler.config import TrainingCompilerConfig  # noqa: F401
