+++ src/sagemaker/huggingface/estimator.py
@@ -26,7 +26,7 @@ from sagemaker.fw_utils import (
 from sagemaker.huggingface.model import HuggingFaceModel
 from sagemaker.vpc_utils import VPC_CONFIG_DEFAULT
 
-from sagemaker.training_compiler.config import TrainingCompilerConfig
+from sagemaker.huggingface.training_compiler.config import TrainingCompilerConfig
 
 logger = logging.getLogger("sagemaker")
 
@@ -199,11 +199,7 @@ class HuggingFace(Framework):
                 )
                 raise ValueError(error_string)
             if compiler_config:
-                compiler_config.validate(
-                    image_uri=image_uri,
-                    instance_type=instance_type,
-                    distribution=distribution,
-                )
+                compiler_config.validate(self)
 
         self.distribution = distribution or {}
         self.compiler_config = compiler_config
